requirements/datetime-month.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --extra=zzz-test-defaults --extra=zzz-test-datetime-month --prerelease=disallow --output-file=requirements/datetime-month.txt --python-version=3.11 pyproject.toml
attrs==23.2.0
    # via hypothesis
coverage==7.6.0
    # via pytest-cov
datetime-month==1.0.6
    # via dycw-utilities (pyproject.toml)
hypothesis==6.106.1
    # via dycw-utilities (pyproject.toml)
iniconfig==2.0.0
    # via pytest
packaging==24.1
    # via
    #   pytest
    #   pytest-rerunfailures
pluggy==1.5.0
    # via pytest
pytest==8.2.2
    # via
    #   dycw-utilities (pyproject.toml)
    #   pytest-cov
    #   pytest-randomly
    #   pytest-rerunfailures
pytest-cov==5.0.0
    # via dycw-utilities (pyproject.toml)
pytest-randomly==3.15.0
    # via dycw-utilities (pyproject.toml)
pytest-rerunfailures==14.0
    # via dycw-utilities (pyproject.toml)
sortedcontainers==2.4.0
    # via hypothesis
typing-extensions==4.12.2
    # via dycw-utilities (pyproject.toml)
